ColorFilterArray: clang-tidy: widening cast

+++ src/librawspeed/metadata/ColorFilterArray.cpp
@@ -63,7 +63,7 @@ CFAColor ColorFilterArray::getColorAt( int x, int y ) const
   x = (x % size.x + size.x) % size.x;
   y = (y % size.y + size.y) % size.y;
 
-  return cfa[x + y * size.x];
+  return cfa[x + (size_t)y * size.x];
 }
 
 void ColorFilterArray::setCFA( iPoint2D in_size, ... )
@@ -91,7 +91,7 @@ void ColorFilterArray::shiftLeft(int n) {
   vector<CFAColor> tmp(size.area());
   for (int y = 0; y < size.y; ++y) {
     for (int x = 0; x < size.x; ++x) {
-      tmp[x + y * size.x] = getColorAt(x+n, y);
+      tmp[x + (size_t)y * size.x] = getColorAt(x + n, y);
     }
   }
   cfa = tmp;
@@ -108,7 +108,7 @@ void ColorFilterArray::shiftDown(int n) {
   vector<CFAColor> tmp(size.area());
   for (int y = 0; y < size.y; ++y) {
     for (int x = 0; x < size.x; ++x) {
-      tmp[x + y * size.x] = getColorAt(x, y+n);
+      tmp[x + (size_t)y * size.x] = getColorAt(x, y + n);
     }
   }
   cfa = tmp;
@@ -173,7 +173,7 @@ void ColorFilterArray::setColorAt(iPoint2D pos, CFAColor c) {
     ThrowRDE("ColorFilterArray::SetColor: position out of CFA pattern");
   if (pos.y >= size.y || pos.y < 0)
     ThrowRDE("ColorFilterArray::SetColor: position out of CFA pattern");
-  cfa[pos.x+pos.y*size.x] = c;
+  cfa[pos.x + (size_t)pos.y * size.x] = c;
 }
 
 static uint32 toDcrawColor( CFAColor c )
